+++ src/Exception/NonDeterministicEncryptorException.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * Copyright (c) Precision Soft
+ */
+
+namespace PrecisionSoft\Doctrine\Encrypt\Exception;
+
+class NonDeterministicEncryptorException extends Exception {}